@@ -1,7 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <Version>1.0.1</Version>
   </PropertyGroup>
 
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
@@ -9,8 +10,8 @@
   </Target>
 
   <ItemGroup>
-    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.3" />
-    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.2" />
+    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.4" />
+    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.3" />
   </ItemGroup>
 
 </Project>
