@@ -1,9 +1,10 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AssemblyName>Smart.Data.Module</AssemblyName>
     <RootNamespace>Smart.Data.Module</RootNamespace>
+    <Version>1.0.1</Version>
   </PropertyGroup>
 
   <ItemGroup>
@@ -12,10 +13,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.3" />
-    <PackageReference Include="ACoreX.Data.Dapper" Version="1.0.3" />
-    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.2" />
-    <PackageReference Include="ACoreX.WebAPI" Version="1.0.0" />
+    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.4" />
+    <PackageReference Include="ACoreX.Data.Dapper" Version="1.0.4" />
+    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.3" />
+    <PackageReference Include="ACoreX.WebAPI" Version="1.0.1" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.2" />
     <PackageReference Include="System.Data.SqlClient" Version="4.7.0" />
   </ItemGroup>
 
