@@ -1,15 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AssemblyName>Smart.Security.Autentication.Module</AssemblyName>
     <RootNamespace>Smart.Security.Autentication</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.3" />
-    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.2" />
-    <PackageReference Include="ACoreX.WebAPI" Version="1.0.0" />
+    <PackageReference Include="ACoreX.Data.Abstractions" Version="1.0.4" />
+    <PackageReference Include="ACoreX.Injector.Abstractions" Version="1.0.3" />
+    <PackageReference Include="ACoreX.WebAPI" Version="1.0.1" />
   </ItemGroup>
 
   <ItemGroup>
