@@ -1,4 +1,9 @@
-﻿using ACoreX.AssemblyLoader;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
+using ACoreX.AssemblyLoader;
 using ACoreX.Authentication.Core;
 using ACoreX.Authentication.JWT;
 using ACoreX.Injector.Abstractions;
@@ -8,10 +13,10 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using System;
-using System.IO;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 
-namespace Smart.Run
+namespace Smart.View.Run
 {
     public class Startup
     {
@@ -25,6 +30,7 @@ namespace Smart.Run
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            //services.AddControllers();
             string libPath = "";
 
             libPath = Path.Combine(new DirectoryInfo(Environment.CurrentDirectory).Parent.FullName, "lib");
@@ -35,18 +41,24 @@ namespace Smart.Run
            .SetCompatibilityVersion(CompatibilityVersion.Version_3_0)
            .LoadModules(builder, libPath)
            .AddControllers();
-
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
-            app.UseMvc();
+            app.UseRouting();
+
+            app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
